+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Unreleased
 ----------
 - Added support for changing an existing order
+- Introduced `TradeStatus::to_order_status` helper method
 - Implemented `Eq` and `Hash` for the various `Id` types